Remove Pkg2, merge its contents into main dir
Also some file renaming to match module names.
