+++ src/Pkg3.jl
@@ -35,7 +35,8 @@ include("../ext/TOML/src/TOML.jl")
 include("../ext/TerminalMenus/src/TerminalMenus.jl")
 
 include("Types.jl")
-include("Pkg2/Pkg2.jl")
+include("Query.jl")
+include("Resolve.jl")
 include("Display.jl")
 include("Operations.jl")
 include("REPLMode.jl")
